requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
pydantic[email]
python-jose
passlib[bcrypt]
